# swagger.json
{
  "swagger": "2.0",
  "info": {
    "version": "1.0.0",
    "description": "API for E-commerce application",
    "title": "E-commerce API"
  },
  "host": "localhost:3200",
  "securityDefinitions": {
    "JWT": {
      "in": "header",
      "name": "Authorization",
      "type": "apiKey"
    }
  },
  "paths": {
    "/api/products": {
      "get": {
        "tags": ["Products"],
        "summary": "Get Products",
        "description": "User will get all products",
        "security": [{ "JWT": {} }],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "type": "array",
              "items": { "$ref": "#/definitions/ProductResponse" }
            }
          },
          "401": {
            "description": "Unauthorized"
          }
        }
      },
      "post": {
        "tags": ["Products"],
        "summary": "Add a new product (JSON)",
        "description": "Create a new product. Provide imageUrl as a string (filename or URL) in JSON.",
        "security": [{ "JWT": {} }],
        "consumes": ["application/json"],
        "produces": ["application/json"],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "description": "Product payload",
            "required": true,
            "schema": { "$ref": "#/definitions/ProductCreateRequest" }
          }
        ],
        "responses": {
          "201": {
            "description": "Product created",
            "schema": { "$ref": "#/definitions/ProductResponse" }
          },
          "400": {
            "description": "Validation error",
            "schema": { "$ref": "#/definitions/ErrorResponse" }
          },
          "500": {
            "description": "Server error",
            "schema": { "$ref": "#/definitions/ErrorResponse" }
          }
        }
      }
    },
...
  },
  "definitions": {
...
    "ErrorResponse": {
      "type": "object",
      "properties": {
        "message": {
          "type": "string",
          "example": "Validation failed: email is required"
        }
      }
    },
    "ProductResponse": {
      "type": "object",
      "properties": {
        "_id": { "type": "string", "example": "64f2b1e2a1234abcd5678ef0" },
        "name": { "type": "string", "example": "iPhone 13" },
        "desc": { "type": "string", "example": "Latest model" },
        "price": { "type": "number", "example": 999.99 },
        "imageUrl": { "type": "string", "example": "2023-07-24T08_31_00.909Ziphone.png" },
        "category": { "type": "string", "example": "Phones" },
        "inStock": { "type": "integer", "example": 10 },
        "size": { "type": "array", "items": { "type": "string" }, "example": ["S","M","L"] },
        "categories": { "type": "array", "items": { "type": "string" } },
        "reviews": { "type": "array", "items": { "type": "string" } }
      }
    },
...
    "ProductCreateRequest": {
      "type": "object",
      "required": ["name", "price"],
      "properties": {
        "name": { "type": "string", "example": "Phone Pro" },
        "price": { "type": "number", "example": 799.99 },
        "description": { "type": "string", "example": "Latest phone" },
        "inStock": { "type": "integer", "example": 10 },
        "sizes": {
          "description": "Array or comma-separated string",
          "oneOf": [
            { "type": "array", "items": { "type": "string" } },
            { "type": "string" }
          ]
        },
        "categories": {
          "description": "Array of category ids or comma-separated string",
          "oneOf": [
            { "type": "array", "items": { "type": "string" } },
            { "type": "string" }
          ]
        },
        "imageUrl": {
          "type": "string",
          "description": "Filename already uploaded or accessible URL (string)",
          "example": "2023-07-24T08_31_00.909Ziphone.png"
        }
      }
    }
  }
}